# lib/api-zod/src/generated/types/usernameStats.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UsernameStats {
  username: string;
  visitCount: number;
  /** @nullable */
  lastVisit: string | null;
}