# lib/api-zod/src/generated/types/visit.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Visit {
  id: number;
  username: string;
  /** @nullable */
  ip?: string | null;
  /** @nullable */
  isp?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  region?: string | null;
  /** @nullable */
  timezone?: string | null;
  /** @nullable */
  org?: string | null;
  /** @nullable */
  ua?: string | null;
  /** @nullable */
  platform?: string | null;
  /** @nullable */
  screenWidth?: number | null;
  /** @nullable */
  screenHeight?: number | null;
  /** @nullable */
  screenDpr?: number | null;
  /** @nullable */
  memoryGb?: number | null;
  /** @nullable */
  cpuCores?: number | null;
  /** @nullable */
  touchPoints?: number | null;
  /** @nullable */
  lang?: string | null;
  /** @nullable */
  batteryLevel?: number | null;
  /** @nullable */
  batteryCharging?: boolean | null;
  /** @nullable */
  networkType?: string | null;
  /** @nullable */
  networkDownlink?: number | null;
  /** @nullable */
  networkRtt?: number | null;
  /** @nullable */
  rawIpinfo?: string | null;
  createdAt: string;
}