# lib/api-zod/src/generated/types/visitInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VisitInput {
  username: string;
  /** @nullable */
  ua?: string | null;
  /** @nullable */
  platform?: string | null;
  /** @nullable */
  screenWidth?: number | null;
  /** @nullable */
  screenHeight?: number | null;
  /** @nullable */
  screenDpr?: number | null;
  /** @nullable */
  memoryGb?: number | null;
  /** @nullable */
  cpuCores?: number | null;
  /** @nullable */
  touchPoints?: number | null;
  /** @nullable */
  lang?: string | null;
  /** @nullable */
  batteryLevel?: number | null;
  /** @nullable */
  batteryCharging?: boolean | null;
  /** @nullable */
  networkType?: string | null;
  /** @nullable */
  networkDownlink?: number | null;
  /** @nullable */
  networkRtt?: number | null;
}